# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RepoPilot Frontend–Backend API",
    "description": "Contract of HTTP APIs that the frontend (`frontend/src/lib/api.ts`, `auth.ts`) calls against the backend. Derived from frontend clients and response TypeScript types.",
    "version": "4.2.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local backend (dev). Vite proxies `/api` and `/auth` from :5173 to this origin."
    }
  ],
...
  "paths": {
...
    "/api/admin/login": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin login",
        "operationId": "adminLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/overview": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin dashboard overview",
        "operationId": "fetchAdminOverview",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Overview",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminOverview"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/sync-tasks": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List knowledge sync/build tasks across repos",
        "operationId": "fetchAdminSyncTasks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "all"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50,
              "minimum": 1,
              "maximum": 200
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sync tasks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSyncTaskList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/sync-failures": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List failed sync/build tasks",
        "operationId": "fetchAdminSyncFailures",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Failures",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSyncFailureList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/integrity": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Per-repo knowledge/FAQ integrity snapshot",
        "operationId": "fetchAdminIntegrity",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Integrity checks",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminIntegrityList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/faq/repos": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List repos with FAQ counts for export",
        "operationId": "fetchAdminFaqRepos",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "FAQ repo options",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminFaqRepoList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/faq/export": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Batch export FAQ for selected repos",
        "operationId": "exportAdminFaq",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminFaqExportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Exported document",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminFaqExportResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/audit-logs": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List admin audit logs",
        "operationId": "fetchAdminAuditLogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuditLogList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List community users (may be empty)",
        "operationId": "fetchAdminUsers",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "GitHub OAuth access token",
        "description": "Token from GitHub OAuth (`Authorization: Bearer <token>`). Frontend stores it in localStorage and sends it via `apiFetch` when present."
      }
    },
...
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid Bearer token. Frontend clears auth and redirects to `/login`.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "properties": {
          "detail": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  },
                  "additionalProperties": true
                }
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          }
        }
      },
...
      "AdminLoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "AdminLoginResponse": {
        "type": "object",
        "required": [
          "token",
          "username",
          "role"
        ],
        "properties": {
          "token": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      },
      "AdminPlatformStats": {
        "type": "object",
        "required": [
          "totalRepos",
          "syncedRepos",
          "failedRepos",
          "knowledgeChunks",
          "memoryEntries",
          "faqEntries",
          "activeUsers",
          "openIssues",
          "syncSuccessRate",
          "lastFullCheck"
        ],
        "properties": {
          "totalRepos": {
            "type": "integer"
          },
          "syncedRepos": {
            "type": "integer"
          },
          "failedRepos": {
            "type": "integer"
          },
          "knowledgeChunks": {
            "type": "integer"
          },
          "memoryEntries": {
            "type": "integer"
          },
          "faqEntries": {
            "type": "integer"
          },
          "activeUsers": {
            "type": "integer"
          },
          "openIssues": {
            "type": "integer"
          },
          "syncSuccessRate": {
            "type": "number"
          },
          "lastFullCheck": {
            "type": "string"
          }
        }
      },
      "AdminHealthTrendPoint": {
        "type": "object",
        "required": [
          "date",
          "success",
          "failed"
        ],
        "properties": {
          "date": {
            "type": "string"
          },
          "success": {
            "type": "integer"
          },
          "failed": {
            "type": "integer"
          }
        }
      },
      "AdminSyncTask": {
        "type": "object",
        "required": [
          "id",
          "repoFullName",
          "status",
          "startedAt",
          "filesSynced",
          "trigger"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "repoId": {
            "type": "string"
          },
          "repoFullName": {
            "type": "string"
          },
          "owner": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "success",
              "running",
              "failed",
              "paused"
            ]
          },
          "startedAt": {
            "type": "string"
          },
          "endedAt": {
            "type": [
              "string",
              "null"
            ]
          },
          "filesSynced": {
            "type": "integer"
          },
          "errorMessage": {
            "type": [
              "string",
              "null"
            ]
          },
          "trigger": {
            "type": "string",
            "enum": [
              "manual",
              "webhook",
              "scheduled"
            ]
          }
        }
      },
      "AdminOverview": {
        "type": "object",
        "required": [
          "stats",
          "healthTrend",
          "recentSyncTasks"
        ],
        "properties": {
          "stats": {
            "$ref": "#/components/schemas/AdminPlatformStats"
          },
          "healthTrend": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminHealthTrendPoint"
            }
          },
          "recentSyncTasks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSyncTask"
            }
          }
        }
      },
      "AdminSyncTaskList": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSyncTask"
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "AdminSyncFailure": {
        "type": "object",
        "required": [
          "id",
          "repoFullName",
          "failedAt",
          "errorType",
          "errorMessage",
          "retryCount",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "repoFullName": {
            "type": "string"
          },
          "failedAt": {
            "type": "string"
          },
          "errorType": {
            "type": "string",
            "enum": [
              "network",
              "auth",
              "rate_limit",
              "webhook",
              "parse"
            ]
          },
          "errorMessage": {
            "type": "string"
          },
          "retryCount": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "retrying",
              "ignored"
            ]
          }
        }
      },
      "AdminSyncFailureList": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSyncFailure"
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "AdminIntegrityCheck": {
        "type": "object",
        "required": [
          "repoFullName",
          "knowledgeOk",
          "memoryOk",
          "faqOk",
          "chunkCount",
          "memoryCount",
          "lastChecked",
          "issues"
        ],
        "properties": {
          "repoId": {
            "type": "string"
          },
          "repoFullName": {
            "type": "string"
          },
          "knowledgeOk": {
            "type": "boolean"
          },
          "memoryOk": {
            "type": "boolean"
          },
          "faqOk": {
            "type": "boolean"
          },
          "chunkCount": {
            "type": "integer"
          },
          "memoryCount": {
            "type": "integer"
          },
          "lastChecked": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AdminIntegrityList": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminIntegrityCheck"
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "AdminFaqRepoOption": {
        "type": "object",
        "required": [
          "repoId",
          "repoFullName",
          "faqCount",
          "memoryCount",
          "lastUpdated"
        ],
        "properties": {
          "repoId": {
            "type": "string"
          },
          "repoFullName": {
            "type": "string"
          },
          "faqCount": {
            "type": "integer"
          },
          "memoryCount": {
            "type": "integer"
          },
          "lastUpdated": {
            "type": "string"
          }
        }
      },
      "AdminFaqRepoList": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminFaqRepoOption"
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "AdminFaqExportRequest": {
        "type": "object",
        "required": [
          "repoIds"
        ],
        "properties": {
          "repoIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "format": {
            "type": "string",
            "enum": [
              "markdown",
              "json"
            ],
            "default": "markdown"
          }
        }
      },
      "AdminFaqExportResponse": {
        "type": "object",
        "required": [
          "format",
          "content",
          "itemCount",
          "repoCount",
          "exportedAt"
        ],
        "properties": {
          "format": {
            "type": "string",
            "enum": [
              "markdown",
              "json"
            ]
          },
          "content": {
            "type": "string"
          },
          "itemCount": {
            "type": "integer"
          },
          "repoCount": {
            "type": "integer"
          },
          "exportedAt": {
            "type": "string"
          }
        }
      },
      "AdminAuditLog": {
        "type": "object",
        "required": [
          "id",
          "admin",
          "action",
          "target",
          "result",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "admin": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "target": {
            "type": "string"
          },
          "result": {
            "type": "string",
            "enum": [
              "success",
              "failed"
            ]
          },
          "createdAt": {
            "type": "string"
          }
        }
      },
      "AdminAuditLogList": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminAuditLog"
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "AdminCommunityUser": {
        "type": "object",
        "required": [
          "id",
          "login",
          "email",
          "boundRepos",
          "status",
          "lastLogin",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "login": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "boundRepos": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended"
            ]
          },
          "lastLogin": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        }
      },
      "AdminUserList": {
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminCommunityUser"
            }
          },
          "total": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}